% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_graph.R
\name{init_graph}
\alias{init_graph}
\title{Initialising a random graph}
\usage{
init_graph(n, deg)
}
\arguments{
\item{n}{Number of nodes in the network}

\item{deg}{Average degree in the network}
}
\value{
Returns an adjacency matrix
}
\description{
Initialising a random graph
}
\examples{
init_graph(n = 10, deg = 4)
}
