% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_bnr.R
\name{iterate_bnr}
\alias{iterate_bnr}
\alias{iterate_bnr,matrix,logical,numeric,numeric,numeric-method}
\alias{iterate_bnr,matrix,missing,numeric,numeric,numeric-method}
\alias{iterate_bnr,matrix,logical,missing,numeric,numeric-method}
\alias{iterate_bnr,matrix,missing,missing,numeric,numeric-method}
\title{Iterating a bnr network}
\usage{
iterate_bnr(adjm, singleparent, pb, pn, pr)

\S4method{iterate_bnr}{matrix,logical,numeric,numeric,numeric}(adjm, singleparent, pb, pn, pr)

\S4method{iterate_bnr}{matrix,missing,numeric,numeric,numeric}(adjm, singleparent, pb, pn, pr)

\S4method{iterate_bnr}{matrix,logical,missing,numeric,numeric}(adjm, singleparent, pb, pn, pr)

\S4method{iterate_bnr}{matrix,missing,missing,numeric,numeric}(adjm, singleparent, pb, pn, pr)
}
\arguments{
\item{adjm}{Adjacency matrix}

\item{singleparent}{Logical. If TRUE the focal individual connects to one
parent with probability pb. If FALSE the individual connects to two parents
each with probability pb. Default is TRUE.}

\item{pb}{Probability to connect to parent. Default is 1.}

\item{pn}{Probability to connect to neighbour of parent(s)}

\item{pr}{Probability to connect to individuals that are not connected to
the parent}
}
\value{
Returns an iterated version of the supplied adjacency matrix.
}
\description{
Iterating a bnr network
}
\examples{
1+1
}
