% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bnr.R
\name{make_bnr}
\alias{make_bnr}
\alias{make_bnr,numeric,numeric,numeric,numeric,numeric-method}
\alias{make_bnr,numeric,numeric,missing,numeric,numeric-method}
\title{Dynamic networks updating for Pb, Pn, Pr networks}
\usage{
make_bnr(n, np, pb, pn, pr)

\S4method{make_bnr}{numeric,numeric,numeric,numeric,numeric}(n, np, pb, pn, pr)

\S4method{make_bnr}{numeric,numeric,missing,numeric,numeric}(n, np, pb, pn, pr)
}
\arguments{
\item{n}{Number of vertices (population size)}

\item{np}{numeric vector setting ids for the newborn (i.e. which individual will be replaced with a new one) and a parent(s). Length 2 or 3. If you want to randomly select an id for the newborn (first value) and parents (second and third value), simply use c(0,0) or c(0,0,0).}

\item{pb}{Probability to connect to parent (default is 1)}

\item{pn}{Probability to connect to neighbour of parent(s)}

\item{pr}{Probability to connect to individuals that are not connected to
the parent}
}
\value{
Function returns an adjacncy-matrix.
}
\description{
This function takes adj.matrix (ADJM), probabilities to connect to parent(s), neighbours, and randoms (PB, PN, PR), the index of the parent (if NULL, default, NPARENT number of individuals are randomly choosen as parent), number of parents (NPARENT, default is 1).
}
\details{
It is important to notw that the
}
\examples{
make_bnr(n = 10, np = c(0,0), pb = 1, pn = .2, pr = .01)
}
\references{
none
}
