% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bnr.R
\name{make_bnr}
\alias{make_bnr}
\title{Dynamic networks updating for Pb, Pn, Pr networks}
\usage{
make_bnr(n, pb, pn, pr, singleparent)
}
\arguments{
\item{n}{Number of vertices (population size)}

\item{pb}{Probability to connect to parent (default is 1)}

\item{pn}{Probability to connect to neighbour of parent(s)}

\item{pr}{Probability to connect to individuals that are not connected to
the parent}

\item{singleparent}{Logical. If TRUE the focal individual connects to one
parent with probability pb. If FALSE the individual connects to two parents
each with probability pb. Default is TRUE.}
}
\value{
Function returns an adjacncy-matrix.
}
\description{
`make_bnr()` returns an adjacency matrix.
}
\details{
This function will ...
}
\examples{
make_bnr(n = 100, pb = 1, pn = .2, pr = .01, singleparent = TRUE)
}
